+++ src/hooks/queries/useGetAllMemberListQuery.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { allMemberApi } from "@/apis/allMemberApi";
 import { QueryKey } from "@/constants/queryKey";
-import { SearchVariantType } from "@/types/entities/search";
+import { SearchVariantType } from "@/types/entities/member";
 
 export default function useGetAllMemberListQuery(
   page: number,
