+++ src/types/entities/common.ts
@@ -1,0 +1,6 @@
+export type MemberVariantType = "ASSOCIATE" | "GUEST";
+
+export type PaginationModelType = {
+  pageSize: number;
+  page: number;
+};
